refactor: Dropdown option 타입 파일 분리
