+++ src/components/common/type.ts
@@ -1,0 +1,4 @@
+export type DropdownOptionType = {
+  label: string;
+  value: string;
+};
