+++ src/components/features/product/api/responseMapper.ts
@@ -1,5 +1,5 @@
-import { CartItemContent, ProductContent } from './api/type';
-import { Cart, Product, ProductCategoryType } from './type';
+import { CartItemContent, ProductContent } from './type';
+import { Cart, Product, ProductCategoryType } from '../type';
 
 export const convertResponseToProduct = ({
   id,
